add a systematic sweep test for 6-bit range queries

+++ bit-structures/src/wasm_wavelet_matrix.rs
@@ -101,9 +101,7 @@ impl WaveletMatrix32 {
         assert!(symbol_ranges.len() % 2 == 0,);
         let symbol_ranges: Vec<_> = symbol_ranges.chunks_exact(2).map(|x| x[0]..x[1]).collect();
         let range = range_lo..range_hi;
-        let counts = self
-            .0
-            .count_symbol_ranges(&symbol_ranges, range, dims as usize);
+        let counts = self.0.count_symbol_ranges(&symbol_ranges, range, dims);
         let obj = js_sys::Object::new();
         let err = "could not set js property";
         js_sys::Reflect::set(&obj, &"counts".into(), &Uint32Array::from(&counts[..])).expect(err);
